Resolved conflicts:
- Updated mReceiver to use std::unique_ptr (from upstream)
- Preserved NTRIP client functionality (from ntrip-client branch)
- Added onDeviceSocketStateChanged connection for NTRIP support

+++ src/core/referencingfeaturelistmodel.cpp
@@ -54,7 +54,7 @@ QModelIndex ReferencingFeatureListModelBase::parent( const QModelIndex &index ) 
 int ReferencingFeatureListModelBase::rowCount( const QModelIndex &parent ) const
 {
   Q_UNUSED( parent )
-  return mEntries.size();
+  return static_cast<int>( mEntries.size() );
 }
 
 int ReferencingFeatureListModelBase::columnCount( const QModelIndex &parent ) const
